using provisions and only clearing database when exiting last session stack

+++ tests/conftest.py
@@ -1,18 +1,21 @@
-from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
-from psycopg2 import sql
-from sqlalchemy import text
+import configparser
+import os
+import pathlib
+import tempfile
+
 import psycopg2
-from lightcurvedb.core.connection import DB
+import pytest
+from click.testing import CliRunner
+from psycopg2 import sql
+from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
+from sqlalchemy import text
+
 from lightcurvedb import __version__ as version
 from lightcurvedb import db_from_config, models
 from lightcurvedb.core.base_model import QLPModel
-import configparser
-import pathlib
-import os
-import tempfile
+from lightcurvedb.core.connection import DB
 
-import pytest
-from click.testing import CliRunner
+from . import provision
 
 
 def _db_connection(database):
@@ -21,7 +24,7 @@ def _db_connection(database):
         user=os.environ["POSTGRES_USER"],
         password=os.environ["POSTGRES_PASSWORD"],
         host=os.environ["HOST"],
-        port=os.environ["PORT"]
+        port=os.environ["PORT"],
     )
     return conn
 
@@ -30,15 +33,11 @@ def _create_testdb(testdb_name):
     postgres_conn = _db_connection("postgres")
     postgres_conn.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
     with postgres_conn.cursor() as cur:
-        drop = (
-            sql
-            .SQL("DROP DATABASE IF EXISTS {}")
-            .format(sql.Identifier(testdb_name))
+        drop = sql.SQL("DROP DATABASE IF EXISTS {}").format(
+            sql.Identifier(testdb_name)
         )
-        create = (
-            sql
-            .SQL("CREATE DATABASE {}")
-            .format(sql.Identifier(testdb_name))
+        create = sql.SQL("CREATE DATABASE {}").format(
+            sql.Identifier(testdb_name)
         )
         cur.execute(drop)
         cur.execute(create)
@@ -59,10 +58,10 @@ def _populate_configuration(testdb_name):
     with open(CONFIG_PATH, "wt") as fout:
         parser.write(fout)
 
-
     db = db_from_config(CONFIG_PATH)
     with db:
         QLPModel.metadata.create_all(db.bind)
+        provision.sync_tess_positions(db)
 
     return CONFIG_PATH
 
@@ -71,10 +70,13 @@ class TestDB(DB):
     def close(self):
         if self.is_active:
             self.session.rollback()
-        for table in reversed(QLPModel.metadata.sorted_tables):
-            q = text(f"TRUNCATE TABLE {table.name} CASCADE")
-            self.session.execute(q)
-        self.session.commit()
+        if self.depth <= 1:
+            for table in reversed(QLPModel.metadata.sorted_tables):
+                if table.name == models.SpacecraftEphemeris.__tablename__:
+                    continue
+                q = text(f"TRUNCATE TABLE {table.name} CASCADE")
+                self.session.execute(q)
+            self.session.commit()
         return super().close()
